+++ PortFolioBackEnd/src/main/java/com/rer/backend/repositorios/ExperienciaLaboralRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExperienciaLaboralRepositorio extends JpaRepository<ExperienciaLaboral, Integer> {   
+public interface ExperienciaLaboralRepositorio extends JpaRepository<ExperienciaLaboral, Long> {   
 }
